HungUp

+++ database/migrations/2024_05_24_100231_create_discount_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name',50);
             $table->datetime('date_start');
             $table->datetime('date_end');
-            $table->decimal('percent',5,2);
+            $table->integer('percent');
             $table->boolean('status')->default(1);
             $table->timestamps();
         });
